+++ scalahost/integration/src/main/scala/example/Sugar.scala
@@ -2,4 +2,5 @@ package example
 
 class Sugar {
   List(1).map(_ + 2)
+  Array.empty[Int].headOption
 }
