+++ scalahost/nsc/src/main/scala/scala/meta/internal/semantic/LengthWriter.scala
@@ -1,0 +1,17 @@
+package scala.meta.internal.semantic
+
+import java.io.Writer
+
+// A writer that keeps track of the current length.
+class LengthWriter(delegate: Writer, start: Int) extends Writer {
+  var length: Int = start
+  override def flush(): Unit = {
+    length = 0
+    delegate.flush()
+  }
+  override def write(cbuf: Array[Char], off: Int, len: Int): Unit = {
+    length += len
+    delegate.write(cbuf, off, len)
+  }
+  override def close(): Unit = delegate.close()
+}
